+++ example/lib/detail_page.dart
@@ -5,31 +5,38 @@ class DetailPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          title: const Text('Detail Page'),
-        ),
-        body: Center(
-            child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            const Text('Detail Page'),
-            TextButton(
-                onPressed: () {
-                  showDialog(
-                    context: context,
-                    builder: (ctx) => AlertDialog(
-                      title: const Text('Dialog'),
-                      actions: <Widget>[
-                        TextButton(
-                            onPressed: () => Navigator.pop(ctx),
-                            child: const Text('OK')),
-                      ],
-                    ),
-                  );
-                },
-                child: const Text('Show Dialog'))
-          ],
-        )));
+    final args = ModalRoute.of(context)!.settings.arguments;
+    return WillPopScope(
+      onWillPop: () async {
+        Navigator.pop(context, 'Pop result from ${this.runtimeType}');
+        return false;
+      },
+      child: Scaffold(
+          appBar: AppBar(
+            title: Text(args.toString()), //?? 'Detail Page'),
+          ),
+          body: Center(
+              child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              const Text('Detail Page'),
+              TextButton(
+                  onPressed: () {
+                    showDialog(
+                      context: context,
+                      builder: (ctx) => AlertDialog(
+                        title: const Text('Dialog'),
+                        actions: <Widget>[
+                          TextButton(
+                              onPressed: () => Navigator.pop(ctx),
+                              child: const Text('OK')),
+                        ],
+                      ),
+                    );
+                  },
+                  child: const Text('Show Dialog'))
+            ],
+          ))),
+    );
   }
 }