+++ kits/flutter_ume_kit_device/lib/components/device_info/device_info_panel.dart
@@ -1,4 +1,3 @@
-import 'dart:convert';
 import 'package:device_info/device_info.dart';
 import 'package:flutter/material.dart';
 import 'package:platform/platform.dart';
@@ -17,8 +16,7 @@ class DeviceInfoPanel extends StatefulWidget implements Pluggable {
   Widget buildWidget(BuildContext? context) => this;
 
   @override
-  ImageProvider<Object> get iconImageProvider =>
-      MemoryImage(base64Decode(icon.iconData));
+  ImageProvider<Object> get iconImageProvider => MemoryImage(icon.iconBytes);
 
   @override
   String get name => 'DeviceInfo';