+++ test/util/store_mixin_test.dart
@@ -1,12 +1,5 @@
-import 'dart:async';
-import 'dart:collection';
-import 'dart:typed_data';
-
-import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:tuple/tuple.dart';
 
 import '../utils/mock_classes.dart';