+++ kits/flutter_ume_kit_ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.0.0
+
+* 正式版
+
+* Normal version.
+
+## 1.0.0-dev.0
+
+* 适配 Flutter 3
+
+* Adapt Flutter 3
+
 ## 0.3.0
 
 * 新增 UI 调试插件：cyclop、touch_indicator